Move initial endpoints into their own objects
